test: Add TestHttpMessageHandler for simulating HTTP responses in OfrepClient tests
