+++ test/OpenFeature.Providers.Ofrep.Test/Client/OfrepClientTest.cs
@@ -1,5 +1,4 @@
 using System.Net;
-using System.Text;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -8,9 +7,10 @@ using OpenFeature.Providers.Ofrep.Client;
 using OpenFeature.Providers.Ofrep.Client.Exceptions;
 using OpenFeature.Providers.Ofrep.Configuration;
 using OpenFeature.Providers.Ofrep.Models;
+using OpenFeature.Providers.Ofrep.Test.Helpers;
 using Xunit;
 
-namespace OpenFeature.Providers.Ofrep.Test;
+namespace OpenFeature.Providers.Ofrep.Test.Client;
 
 public class OfrepClientTest : IDisposable
 {
@@ -1023,75 +1023,6 @@ public class OfrepClientTest : IDisposable
 
         // Act & Assert
         Assert.Throws<ArgumentOutOfRangeException>(() => client.SetCacheDuration(slightlyOverOneDay));
-    }
-
-    #endregion
-
-    #region Helper Classes
-
-    private class TestHttpMessageHandler : HttpMessageHandler
-    {
-        private readonly Queue<(HttpStatusCode statusCode, string content, string? etag, Exception? exception)>
-            _responses = new();
-
-        private readonly List<HttpRequestMessage> _requests = new();
-
-        public IReadOnlyList<HttpRequestMessage> Requests => this._requests.AsReadOnly();
-
-        public void SetupResponse(HttpStatusCode statusCode, string content, string? etag = null)
-        {
-            this._responses.Enqueue((statusCode, content, etag, null));
-        }
-
-        public void SetupException(Exception exception)
-        {
-            this._responses.Enqueue((HttpStatusCode.OK, "", null, exception));
-        }
-
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
-            CancellationToken cancellationToken)
-        {
-            this._requests.Add(request);
-
-            if (!this._responses.TryDequeue(out var response))
-            {
-                response = (HttpStatusCode.OK, "{}", null, null);
-            }
-
-            if (response.exception != null)
-            {
-                throw response.exception;
-            }
-
-            var httpResponse = new HttpResponseMessage(response.statusCode)
-            {
-                Content = new StringContent(response.content, Encoding.UTF8, "application/json")
-            };
-
-            if (!string.IsNullOrEmpty(response.etag))
-            {
-                httpResponse.Headers.ETag = new System.Net.Http.Headers.EntityTagHeaderValue(response.etag);
-            }
-
-            // If status code indicates an error and no specific exception was set,
-            // let EnsureSuccessStatusCode handle it
-            return Task.FromResult(httpResponse);
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                foreach (var request in this._requests)
-                {
-                    request.Dispose();
-                }
-
-                this._requests.Clear();
-            }
-
-            base.Dispose(disposing);
-        }
     }
 
     #endregion
